+++ socat-emitter-wsl.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+powershell.exe -c Get-Clipboard | socat - tcp:localhost:8122